Created component to generate access token
Created an api call to Reddit's api to generate an access token
to authenticate a user. This will allow a user to generate a
unique access token so they can authenticate the app to have access
to their account.

Next step is to pass the access token into the the api and get
access to a user's account. I also need to refactor the code to
be cleaner and easier to read so as the app scales, the code doesn't
become too confusing.