Setup the first stap of user authentication
Set up the UserCredentials component which redirects the user
to a reddit authorization page. When the user grants permission to
reddit for the app to act on their behalf.

This is a refactored version of how this was achieved before to be
more secure.

Next step is to pass along authorization code to the api to make
a call on the user's behalf.