Implemented OAuth 2.0 authentication
Setup the functionality so that a user can choose to allow the
app to have access to their profile information. Previously,
the user would pass their username and password to authenticate
the app, however this is not considered secure.

The new means of authentication is much more secure for both the app
and the user.