Cleaned up components and added a landing page
In addition to cleaning up the components handling the
2.0 Oauth flow, I also added a landing page so that
when a user authenticates the app to have access to
their user history, the app will automatically
redirect them to a landing page where we can pass
the authorization token into the reddit api
fetch request.

I also seperated the data and logic for each component
and cleaned up some of the naming conventions to be
clearer. Finally I added some conditional checks to
prevent unwanted behavior throughout the app.