Added means of fetching refresh token for additional api requests
Previously the user would have to authorize the app each time
they wanted access to their user information. Now the user
can utilize a refresh token to make additional api requests
to different endpoints of the reddit API.

I also refactored code to be clearer on what each function
is doing. Finally I removed the button to authorize the app
and now the app is immediately authorized after the user
allows the app to have access to their user info and after
the app redirects the user back to a landing page. This way
the app automatically fetches their user info and generates
a refresh token behind the scenes.