Added automatic rendering of user's profile data
Whenever a user authorizes the app to have access to their account
the landing page automatically renders the user's profile information.

Currently the process is seamless. The next step is to establish
a local server that stores the user's profile info so that if the
page is refreshed, the user doesn't have to make another API request
to grab the user profile data.