Added loading state and refactored landing page code
I moved the api calls and fetch requests for the user profile data
inside of a useEffect so that it only makes calls to the api when
an authToken or a refreshToken is returned.

I also added a loading state so we can have a loading animation
in between fetch calls. I also added an element so either no content
or a special message will render if there isn't anything populating
the userProfile prop passed down to the landing component.