User profile persists between page refreshes
After a user authenticates the app to have access to their profile
data, the app can now store that data in the browser's cache so it
can be retrieved in between page refreshes. This prevents the user
from having to make calls to the api for profile data after each
page refresh.

This was also the most appropriate means of storing data as we're
not returning any confidential information.