Refresh tokens now persist across all components
Currently I'm utilizing sessionStorage to store the refresh token
and have it persists across the entire app so it can be used to make
api calls for additional components.

However, this is not considered a secure means of protecting the
web application against cross site attacks. Will look into additional
more secure means of storing the refresh token to persist across all
components.