Added a nav bar component to work with for now
When you click on different aspects of the nav bar it'll display
specific components that will be used for visualizing data. The
styling is minimal for now and I need to fix a bug where the user
profile card rerenders for a split second in between in clicking
on components.