Fixed bug making unnecessary api calls for data
Previously if a user did not refresh or cache a page for any means
and tried navigating to a different component in the nav bar, the
app would navigate to a different page but would error out because
it was trying to make another api call for user data.

This was unintended behavior and was fixed to only make api calls
after the user authenticates the app.