Able fetch user comments from Reddit API endpoint
Also implemented a custom hook and username context so that the
whole application can share the username state. This makes it easier
to make API calls using the username state to endpoints that require
the username state.

Also stored the username in session storage so that we can get
the username state after subsequent page refreshes.