Added styling to the Subscribed Subreddit section
Also refactored bits of code to be a bit cleaner in their
functionality, readability and usage. Overall I edited styling
on most of the components so that the site looks a bit more
presentable overall. I also added in custom colors for borders,
text, animations and etc.

Next step is to alter the styling on Posts and Comments to make things
across the site look consistent.