Introduced an expand button to the posts section
Currently the component will render div elements for each post. When
the post renders each div element any post with selftext that exceeds
500 characters has a "See more..." button attached to it. If that
button is pressed, a function passes the post ID (initially retrieved
from the returned JSON) to an isExpanded variable init. to the
expanded state (which is an object) with the postId as a value of true.

We're also spreading prevExpanded states to the expanded state so that
as each element opens, it's postId key/value is appended to the state
object allowing for future posts to also be appended. This prevents
other posts from closing when we expand one post.