Styled comments & refactored div structures
In addition to doing a lot of the styling for the comments section,
I also refactored for the Posts section as well.

I tried to make the layout and style to look more like Reddit's site.
Next step is to add an expand option to any comments over 500
characters.

I also want to add some comment or post flairs as well as a possible
comment icon. Then I want to go back over the web app one more time
to make sure everything is consistent. I also want to add a planned
feature list for the welcome page.