Abstracted expand/hide features into a component
Rather than repeating the code for expanding and hiding content
over 500 characters, I abstracted the functionality into a
component. This prevents the developer from having to read two
separate pieces of code that do the same thing and cuts down on the
amount of code needed for both the comment and posts components.