Created component to parse through text for hyperlinks
Utilized the React-Linkify component which parses through a body
of text and finds URLs. The URLs are converted to the text
"Redirect Link..." along with classNames passed to it. The next
step is to address links that are repeated twice in a row possibly,
however this might not be worth the time.

The next step is to also address media posted in the posts that are
not links, but may appear as links.