Display domains from URLs rather than full URL
In the comments body component, I modified the URL parse component
to extract the domain from the URL and to only display the domain,
rather than the full url to help shorten the amount of content
necessary to display.

The next step is to modify the comment body to better be contained.
Currently any content that exceeds the comment body spills out of
the comment body and doesn't look very good.