Embedded images in post body
Created a new component that parses through text and finds images
embedded inside of a post body which is then rendered as an embedded
image. The component was applied to both the comments and posts
components.

I also removed the width and height requirements from the
authentication page so it remains responsive.

The next step is to accommodate for embedded videos and then GIFs
before I move on to integrating data visualization into the app.