Added logic to clean up URL ParseUrl
Added logic to grab reddit preview URLs to later remove. They tend
to render when you're trying to grab any images embedded in posts.
This cleans up the design above images, however it will still display
'()' when a user adds captions to their images. Next step is to add
logic to address this.