Added a test bar graph for D3 integration
Started adding in the D3 dependency to start integrating data
visualization into the app. Currently D3 is only visualizing static
data that is not tied to any user information. This was meant to be
an exercise to learn more about D3 and how it works.

The next step is to abstract the bar graph code into it's own component,
make an api call to Reddit with the access token and decide what data
we want to grab and visualize as well as how we want to display it.