Created a bar graph to render post statistics
I finished swapping out the dummy data for the bar graph to render
the post statistics rather than fake data. I also changed the bar graph
to be more consistent with reddit's branding.

I also abstracted out the D3 components into a D3 page and a D3
component. The final thing we need to do is integrate a Giphy API
to render embedded gifs, make the site responsive and fix the tick
formatting on the bar graph.