Integrated a GIF parser for embedded GIFs
I created a GIFParser component that parses through returned
objects. It parses through the GIF object and finds the returned
ID from the GIF object. After it grabs the ID, it then renders the
GIF embedded in each post.

The next step is to format the comments section so it's clear comments
are separated between each. Currently they show up all in one place
and it's hard to tell which are new and which are old comments from
a comment thread.

I also need to format the posts to be cleaner so content doesn't look
as massive.