Adjusted content styling and added graph label
In addition to putting a label on the bar graph, I adjusted the page
container padding for the content body of all the pages. I need to
make everything responsive, but it gives me a general idea of how
everything is going to look.

Might add another graph for comments and for different metrics, but
will see where I end up at the end of the week.