Swapped out iframe el for video el in gif parser
In addition to making further tweaks to how the site is rendered
on mobile devices, I also swapped out the iframe element being used
in the GIF parser for the HTML video element.

This prevents an issue / error from rendering where the site spits
back out an "issue" that the browser can't determine whether to send
a cookie in a cross-site request.

Using the video element also renders the gifs quicker and cuts down
on the perfomance load.